Removed the "no translation" from album.description area.
renamed test files to working file names. Removed old files.

Now we have refactored albums into an album component.
