+++ frontend/src/App.tsx
@@ -11,8 +11,7 @@ import './App.css';
 import MediaComponent from './Contexts/Homepage/MediaComponent';
 import Footer from './Components/Footer/Footer';
 import Visuals from "./Views/Visuals/Visuals";
-// import Promo from "./Views/Promo/Promo";
-import Promo from "./Views/Promo/PromoTest";
+import Promo from "./Views/Promo/Promo";
 import LandingPage from "./Views/Homepage/LandingPage";
 import Videos from "./Views/Videos/Videos";
