--- toDo.txt
@@ -1,13 +1,0 @@
-Currently working to make Albums class like bandMembers
-
-Using
-PromoTest.BandListTextStyle
-Albums.tsx
-
-based off of
-BandMemberList.tsx
-About.tsx
-Promo.tsx
-AlbumOnePromo.tsx
-AlbumTwoPromo.tsx
-
